ui-streamlit/requirements.txt
streamlit
psycopg2-binary
bcrypt
sqlalchemy
pandas
matplotlib
seaborn
streamlit-tags